Add missing imports from the request builders refactor
