+++ src/tcgwars/logic/groovy/TcgStatics.groovy
@@ -5,6 +5,7 @@ import tcgwars.logic.effect.ability.Ability.ActivationReason
 import tcgwars.logic.effect.blocking.*
 import tcgwars.logic.effect.getter.*
 import tcgwars.logic.effect.gm.*
+import tcgwars.logic.client.requestbuilder.*
 import tcgwars.logic.client.*
 import tcgwars.logic.exception.EffectRequirementException
 import tcgwars.logic.util.*
